requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
Django==4.2.6
django-environ==0.11.2
django4-background-tasks==1.2.9
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
lxml==4.9.3
numpy==1.26.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
